Added pyhackernews.pu
This was just me spending some time to learn XPath and lxml.html a bit.
It works fairly well, I thought that I would post it here.

Note: By 'fairly well,' I mean that I tested most of the pieces of code
here separately in ipython before adding them to the actual file, but I
haven't bothered to test the entire thing on concert, so there are
probably kinks that need to be worked out.